backend/auth_service/requirements.txt
fastapi[all]
aiohttp
sqlalchemy
asyncpg
sqlalchemy-utils
redis
pyjwt
faststream
aio-pika
bcrypt